Top 10 Agile Frameworks.md
What is an Agile Framework?
In short, an Agile framework is the set of tools a product manager has for getting better all the time and adapting to change adeptly. If you’re on the hunt for overnight success or easy shortcuts, this isn’t your stop. Agile means committing to the process through thick and thin.

Agile frameworks aren’t just rulebooks you blindly follow. Think of them as a specific approach inspired by that big idea—the Agile Manifesto.